chore: refactor resource read
