+++ ckanext/custom_schema/fiscal/templates/scheming/package/snippets/additional_info.html
@@ -1,14 +1,14 @@
 {% extends "package/snippets/additional_info.html" %}
 
 {%- set exclude_fields = [
-    'id',
-    'title',
-    'name',
-    'notes',
-    'tag_string',
-    'license_id',
-    'owner_org',
-    ] -%}
+  'id',
+  'title',
+  'name',
+  'notes',
+  'tag_string',
+  'license_id',
+  'owner_org',
+  ] -%}
 
 {% block package_additional_info %}
   {%- for field in schema.dataset_fields -%}
