+++ ckanext/custom_schema/nasalsda/templates/scheming/package/resource_read.html
@@ -1,39 +1,40 @@
 {% extends "package/resource_read.html" %}
 
 {%- set exclude_fields = [
-    'name',
-    'description',
-    'url',
-    'format',
-    ] -%}
+  'name',
+  'description',
+  'url',
+  'format',
+  ] -%}
 {%- set schema = h.scheming_get_dataset_schema(dataset_type) -%}
 
 {% block resource_additional_information_inner %}
   {% if res.datastore_active %}
     {% set ddict=h.datastore_dictionary(res.id) %}
-    <div class="module-content">
-      <h2>{{ _('Data Dictionary') }}</h2>
-      <table class="table table-striped table-bordered table-condensed" data-module="table-toggle-more">
-        <thead>
-          <tr>
-            <th scope="col">{{ _('Column') }}</th>
-            <th scope="col">{{ _('Type') }}</th>
-            <th scope="col">{{ _('Label') }}</th>
-            <th scope="col">{{ _('Description') }}</th>
-          </tr>
-        </thead>
-        {% for f in ddict %}
-          <tr {% if loop.index > 5 %}class="toggle-more"{% endif %}>
-            <td>{{ f.id }}</td>
-            <td>{{ f.type }}</td>
-            <td>{{ f.get('info', {}).label }}</td>
-            <td>{{ h.render_markdown(f.get('info', {}).notes) }}</td>
-          </tr>
-        {% endfor %}
-      </table>
-    </div>
+    {% if h.og_is_data_dict_active(ddict) %}
+      <div class="module-content">
+        <h2>{{ _('Data Dictionary') }}</h2>
+        <table class="table table-striped table-bordered table-condensed" data-module="table-toggle-more">
+          <thead>
+            <tr>
+              <th scope="col">{{ _('Column') }}</th>
+              <th scope="col">{{ _('Type') }}</th>
+              <th scope="col">{{ _('Label') }}</th>
+              <th scope="col">{{ _('Description') }}</th>
+            </tr>
+          </thead>
+          {% for field in ddict %}
+            <tr {% if loop.index > 10 %}class="toggle-more"{% endif %}>
+              <td>{{ field.id }}</td>
+              <td>{{ field.type }}</td>
+              <td>{{ field.get('info', {}).label }}</td>
+              <td>{{ h.render_markdown(field.get('info', {}).notes) }}</td>
+            </tr>
+          {% endfor %}
+        </table>
+      </div>
+    {% endif %}
   {% endif %}
-
   <div class="module-content">
     <h2>{{ _('Additional Information') }}</h2>
     <table class="table table-striped table-bordered table-condensed" data-module="table-toggle-more">
@@ -68,14 +69,12 @@
             <td>{{ res.mimetype_inner or res.mimetype or res.format or _('unknown') }}</td>
           </tr>
         {%- endblock -%}
-{#
-        {%- block resource_license -%}
+        {% if res.size and res.size|int != 0 %}
           <tr>
-            <th scope="row">{{ _('License') }}</th>
-            <td>{% snippet "snippets/license.html", pkg_dict=pkg, text_only=True %}</td>
+            <th scope="row">{{ _('Size') }}</th>
+            <td>{{ h.localised_filesize(res.size) }}</td>
           </tr>
-        {%- endblock -%}
-#}
+        {% endif %}
         {%- block resource_fields -%}
           {%- for field in schema.resource_fields -%}
             {%- if field.field_name not in exclude_fields
